TextInputWidget: Update comment about Blink height miscalculation
The original issue from f1dd60e18 (T73310) seems to no longer be present,
but this is still necessary to prevent silly scrollbars when zoomed in.
In my testing, measurementError equals 0.8571420000000103 when zoom level
is 175%, resulting in a scrollbar without this hack. (It takes different
values between -1 and 1 on different zoom levels, most however do not
result in spurious scrollbars.)

Change-Id: I12404c0e9f075d9a331e5840c7ebcf4f38fafa2c