LabelElement: Cast label to string before check if it is empty
An object casts to boolean true, no matter if the string content is
empty or not. That means, that a check for a non-empty label on an
OOUI\HtmlSnippet element will always be true, even if the content
of the label is empty. This results in the addition of the labelElement
class (oo-ui-labelElement) even if the label is empty, which is
not the expected behaviour. The same occurs for the JS implementation.

This change changes the php/js implementation of the LabelElement mixin
to cast to a string value before doing the "is empty" check.

Bug: T129822
Change-Id: I876a1eda3c793a1b100a372dae0b01e764569dfe