InputWidget: Actually reuse parts of the DOM when infusing
It is expected that almost any subclass of InputWidget will override
the #getInputElement method to display its own input. Therefore, if I
rely on the InputWidget implementation of it, it won't work. What the
hell was I thinking.

Moved checking for config.$input into the constructor, before
#getInputElement is called. For subclasses that need to tweak the
original $input, moved that code to constructor also.

Bug: T114408
Change-Id: I3a464fe730dc9382b9cb5936edc21f9bf9b3ed2e