build: Drop the 'npm prepublish' task which runs pre-install as well
Despite the name, `npm prepublish` runs pre-install too, not just -publish.
Upstream refuse to consider this an issue (see npm/npm #3059 on GitHub),
for complicated reasons that mere mortals such as myself cannot hope to
understand, apparently (aka, CoffeeScript is worth helping out at the cost
of pain for others).

This is adding significantly to the build time of every CI operation, as
the CI servers are (relatively) slow and over-burdened, and the full build
operation is lengthy. All `npm install` operations are immediately followed
by `npm test` which runs grunt build in that context, so the only workflow
broken is the actual publication step, for which the one-a-week person
doing it manually is an acceptable cost.

This avoids a lot of useless build time (locally, about a minute; on the
CI cluster's machines, about four minutes) per run of unit tests, which
occur several dozen times a week.

Change-Id: I1e190f494588edb37630861dde48064257d7435e