Prevent modal windows from exceeding available height on Firefox
I feel pretty bad using calc() for anything, but it seems to do the job
here and the 'max-height: 100%' fallback will do fine for browsers that
don't support it. See the task for details.

The `calc(100% - 2em)` is wrapped in `e( '...' )` because otherwise Less
attempted to calculate 100% - 2em itself, deciding it's equal to 98%.

Also added a long medium ProcessDialog to the demo page.

Bug: T131711
Change-Id: Ia2896c7629382391f72152d2065fc956d2d96ecc