ClippableElement: Gracefully handle failure to call clip() after natural height change
By adding a max-height in non-clipped mode that prevents the
Clippable from growing past the point where we would clip.

This is essential for handling cases where something inside a
ClippableElement transitions/animates its height. In the right
circumstances, that can cause clipping to be needed halfway through,
so telling the upstream to call clip() at the right time is infeasible.

Calling clip() for height changes is still required (for example,
a clipped element won't automatically unclip if its height decreases),
but this makes things behave more sensibly if you don't or can't.

Bug: T131056
Change-Id: I2ec0ba74e66fa3f823dcd028a3ce95b050420166