ButtonOptionWidget: Inherit OptionWidget, not DecoratedOptionWidget
DecoratedOptionWidget comes with some styles meant mostly for dropdowns
which are a bit of a pain for buttons. We had tons of CSS to override
them. Instead, inherit just OptionWidget and mix in IconElement and
IndicatorElement ourselves.

This actually fixes a small UI bug in Apex where icons/indicators of
disabled ButtonOptionWidget would be lighter than in any other buttons.

Change-Id: I47a31eb58fcc91124184442b23ef56c00bdc3f92