NumberInputWidget: Disable onWheel action unless the widget has focus
The scroll wheel handling has been annoying me for a long time, it's
very easy to trigger this by accident when scrolling through a form.
I've just realized that Chrome's <input type=number> also has this
feature, but it only kicks in when the field has focus, which seems
like a much better idea.

Change-Id: I6c9f4679dded73da64aeaa91238251a9aee5b40b