ButtonWidget: Implement, document and demonstrate the 'active' config option
ButtonWidget can now be marked as active (`active: true`) to indicate
that clicking the button will only refresh the current page. This is
meant to be used in navigation menus shaped like buttons in PHP. The
JavaScript ButtonWidget was also updated to match, though.

PHP ButtonWidget has gained two new methods: setActive() and isActive().
In JavaScript, they were already implemented on ButtonElement, because
they are used in implementation details of ButtonOptionWidget and
ToggleButtonWidget (and because moving them elsewhere would break
backwards compatibility). Marked them as @protected on ButtonElement
now, and @public on ButtonWidget.

Change-Id: I088c2cc1cbd498046fad4238af1fdad0f3d7467f