MediaWiki theme: Custom `:focus` state for SelectWidgets
We now display focussed SelectWidgets as if the focus was actually on
the selected OptionWidget. This looks more consistent with how native
controls behave (e.g. <input type=radio>).

After this change, focussing a SelectWidget that has nothing selected
will select the first option. The focus states could be implemented
without this behavior change, but it would require quite a bit more work.

* ButtonSelectWidget: Done.
* RadioSelectWidget: Done.
* MenuSelectWidget: No changes needed, can't have focus.
* TabSelectWidget: Needs design (or we might leave it be).
* OutlineSelectWidget: Needs design (or we might leave it be).

Bug: T133429
Change-Id: I52a71fd6d3a9f33d22f5ab1927307384356c003d