MediaWiki theme: Align styles of normal and not-supported SelectFileWidgets
The not-supported style of SelectFileWidget is now aligned with how the
widget appears when the browser does support the feature, in both the
single-line and drop-area forms of the widget.

This also takes the opportunity to hide the dropLabel of disabled
widgets, as urging the user to "drag a file here" isn't helpful when
they cannot.

Bug: T133245
Change-Id: I00a8877fdf38d684fbad051c517b3498c3f77b97