[DEPRECATING CHANGE] CapsuleMultiSelectWidget: Rename to CapsuleMultiselectWidget
CapsuleMultiselectWidget is not a SelectWidget. If a MultiselectWidget
existed, it would not be a subclass of SelectWidget either.

We try to use another class's name as a suffix only when subclassing
things (e.g. Widget–ButtonWidget, InputWidget–TextInputWidget,
SelectWidget–MenuSelectWidget). We actively avoid doing it when not
subclassing (e.g. SelectFileWidget, not FileSelectWidget).

Old name is left as an alias for backwards-compatibility.

Change-Id: I52d243a670accd063b42d43c77eb75158a35c4fe