Implement MultiselectWidget, CheckboxMultiselectWidget and CheckboxMultiselectInputWidget
* MultiselectWidget is an abstract class representing a widget which
  allows the user to select multiple items from a list of predefined
  options.

* CheckboxMultiselectWidget is an implementation of it which uses a
  set of checkboxes for the interface.

* CheckboxMultiselectInputWidget is a wrapper around it which allows
  the selected checkbox values to be submitted with an HTML form.
  It's also available in the PHP version of OOUI.

Stylesheets and the PHP code are adapted from RadioSelectWidget and
RadioSelectInputWidget.

I am quite unhappy how many new classes and how much duplicated code
it took to implement this, but I did not find a sensible and
consistent way to do it more simply. Hopefully this will pay off at
least a little when we make CapsuleMultiselectWidget conform to the
MultiselectWidget interface (T108489).

SelectWidget/OptionWidget and MultiselectWidget/MultioptionWidget are
pretty much unrelated. SelectWidget is coded with only a single
selected option in mind and it didn't make sense to reuse it.

MultioptionWidget is a very silly name.

Bug: T117782
Change-Id: I6f9fad0a462b43ac45c117cd3a3e11c36781cf11