SelectWidget: Improve focus behavior
If possible, only highlight an item when the widget (or one of the
options) gets focus, rather than select it. This matches the behavior
of arrow keys navigation.

Note that this is purely theoretical, as right now as far as I know
the only OptionWidget with 'highlightable' set is MenuOptionWidget,
and MenuSelectWidgets can't have focus anyway (they depend on being
attached to some other widget).

Change-Id: I456e041cf483458af49d40e1a05a5c92eb26becb