Fix GroupElement#removeItems to actually unbind events
The code was missing a loop to set itemEvent. Also rename
variables so the loops in addItems and removeItems match up.

This fixes a bug in DraggableGroupElement, but also add some
extra checks in there so we don't (un)bind events if there is
no change in the current item.

Change-Id: I3221fd0b73a05c41e230c010166f561970119c43