CheckboxMultiselectWidget: Rewrite Shift-clicking code
Just manually call CheckboxMultioptionWidget#setSelected on the items,
rather than trigger event handlers and go through dozens of function calls
before finally ending up in #setSelected anyway.

We can assume that the DOM order of the items and the order of the .items
array is the same; otherwise we'd have to keep some kind of map.

At least on Chrome, the order in which events happen is different for clicks
on the <input> and on the <label>, there are additional fake clicks fired
for non-click actions that change the checkboxes, and e.preventDefault()
doesn't work for some of them. We need some special code to handle this
sanely, but it's not so bad.

Change-Id: I87a1abaf566fe3d7f7a62119cb607697c279494a