build: Add 'prep-test' task to be run before running tests in the browser
Currently the way to get tests results showing in 'tests/index.php' is to
run 'grunt test' first. This does the full build and karma tests and thus
takes a lot of time.

This task does the bare minimum required to get updated tests results
on that interface and takes much lesser time.

Change-Id: I6d6da933320ea33b3daf87f7c870a5c57d07c178