Window: Update `-content` CSS so that child elements can give it focus
Since the elements in oo-ui-window-content are absolutely positioned,
the element itself has no height. Due to this Firefox doesn't give the
parent focus when a child element is clicked.

By setting 'position' to anything other than 'static' we ensure
that the parent gets focus, and 'Esc' to close continues to work.

Bug: T137134
Change-Id: I2806431123ee94b6f8d5f53d6b5afd16aa5c30f9