CapsuleMultiselectWidget: Fix crash on right-click when no input
We are defining this.$input as either null or an input, and yet
on focus/unfocus operations there's a call for resize which
assumes this.$input is a jQuery $( '<input>' ) and tries to manipulate
and calculate data from it.

In general we should allow for an $input in the capsule even if
there's a popup (and have the filtering option as an API the dev
will implement, similar to having to implement the API to add items
from the popup instead of the menu) -- but while we don't allow
for that, any operation on the $input should be guarded.

Change-Id: I464c0ac2a2f6b0129e076037212dfc672e679a4c