FieldsetLayout: Temporarily remove use of `<legend>` due to Chrome 55 bug
Chrome 55 has a major regression where `<legend>` tags styled with
'float' disappear, and those styled with `display: inline-block` etc.
appear to the side of the parent instead of on top. Both bugs affect
us, and the buggy version just recently got promoted into stable.

https://bugs.chromium.org/p/chromium/issues/detail?id=661230
https://bugs.chromium.org/p/chromium/issues/detail?id=670837

I intend to revert this as soon as the Chrome bug is fixed.

Bug: T151061
Change-Id: I73f95965694ec7fb0fa9a474742286e1105e5c85