FieldLayout: Move <label> from $body to $label
When the <label> element (currently, $body is one) wraps anything else
than the actual label and the actual input it pertains to, weird
things happen. In particular, for ActionFieldLayout, hovering/clicking
the button causes hover/click effects on the main input.

We put <label> on the $body, because we were relying on the nesting for
the <label> to work. But since I3424ecfd741b5e268513a5dd778e5082de8f7674
added the 'for' attribute, we no longer need this.

Additionally, let's always use <label> rather than <div> for consistency.
We did this to avoid wrapping <label> around complicated widgets, but
now we no longer wrap it around anything.

Change-Id: Ic4651a7d0868c505dee85ace8d7ac3b6017cfac1