docs: Document for JSDuck various overridden inherited properties
When a property is not annotated with /** @inheritdoc */, it will
not be taken into account by JSDuck when generating documentation.
This resulted in some incorrect information, e.g. MessageDialog
documentation mentioned a static 'name' property that defaults to '',
inherited from Dialog, even though MessageDialog sets it to 'message'.

Additionally, we need to specify /** @static */ for inherited static
properties, otherwise JSDuck assumes they are instance properties.

Also added documentation for some static properties of OptionWidget
and ToolGroup.

Change-Id: I8f5420d28f81b78cbfc1eca8cc71eb0755c112cc