WindowManager: Clarify #addWindows documentation
Everything I write here (except for declaring that using the array
version is "recommended") was technically already documented, but
without knowing a lot of context one couldn't really understand it.
Hopefully this helps.

The code example is not an executable @example because apparently
that doesn't work in method documentation. Grumble.

Change-Id: I7d9d26c4a8f1974db2aea6d5bbb277c8299811fc