ClippableElement: Order matters (inexplicably)
In Chrome, if `overflow-y` is not set before `height`, unnecessary
scrollbars are displayed.

This is incredibly silly, but feels like a less heavy-handed solution
than calling `reconsiderScrollbars()` (our generic workaround for
scrollbars being messed up). And unlike in the case where we stop
clipping, here "just" doing it in this order and forcing a reflow is
sufficient.

Bug: T157672
Change-Id: Id1e0ceb89557db597274d5464289118137bf267b