PopupWidget (and similar): Document why it is unwise to show unattached widgets, and emit warnings
This should help prevent future problems similar to T158141.

A deprecation warning is emitted once per widget. The ones in
ClippableElement and FloatableElement would be sufficient, but
PopupWidget and MenuSelectWidget have extra ones to avoid confusing
theirs users with the clippable/floatable internals.

In the future, the deprecation warnings should be changed to explicitly
throwing an exception. But since this may be happening in uncommon
code paths, we should probably keep them for a few releases.

Change-Id: I4153635b1d8853ca460a526cbfe63a98bd102235