TextInputWidget: Use .prop() rather than .attr() for 'required'
jQuery Migrate erroneously complains about this:
  jQuery.fn.removeAttr no longer sets boolean properties: required

But in fact, the property is automatically bidirectionally synchronized
with the attribute, this code works just fine with jQuery 3, that
message is therefore wrong, and this change is completely unnecessary
other than to appease jQuery Migrate.

Effectively, this changes the attribute we set from `required="required"`
to `required=""`. The getDomElementSummary() function used in our tests
is updated to handle this change. This could be a breaking change for
CSS using the `[required='required']` selector (although such code
should have used just `[required]`, which will still work).

Bug: T160802
Change-Id: I8179eb3f1cef1996f2d837d1493a5217ce4171b3