MenuSelectWidget: Highlight the first result when searching
When we search for an item, highlight the first result found.
This is similar to what happens when we move with the arrow keys
and makes sense for highlighting the first search. It also gives
the option to the user of this widget to call on
this.getHighligtedItem() to pick the first result of the search.
Similar also to an autocomplete feature.

This will be done optionally through the highlightOnFilter
config option.

Change-Id: I66339c2efaf9b7b9a3ea88d59ad087ae102dbbad