ClippableElement: Fix progressive width loss bug
When a ClippableElement is not clipping, has its idealWidth set
to N pixels, and also has a clippable container ancestor (not equal
to the clippable element) whose width is constrained to N pixels
plus K pixels for the border, then on every call to clip(), the
width of the clippable element would be reduced by K pixels.

In practice, K would often be 2 (1px of border on each side),
and the width of the clipped element would shrink by 2px
every time you scrolled the page.

Fix this by not subtracting extraWidth from idealWidth; it doesn't
make sense to do that. Update the setIdealSize() documentation to
clarify this.

Change-Id: I978d80da2e3720efcf07402acb5705f013502b4a