MenuOptionWidget: Remove theme-independent 'check' icon
A 'check' icon was added to any MenuOptionWidgets independent of the
chosen theme, but it was only shown in Apex. Instead removing it per
default and adding it back in Apex through a CSS hack.

The hack has to replicate a bunch of CSS rules from IconElement and
DecoratedOptionWidget. It is needed because while an icon is added
per CSS only the parent element does not get the `.oo-ui-iconElement`
class.

Also removing the CSS workaround in MediaWiki theme to hide default,
unwanted 'check' icons in the MenuOptionWidget (T101560) and adding demo.

Bug: T87835
Bug: T101560
Change-Id: If34ab5302767b8505650a079b033af5506e5cb27