TagMultiselectWidget: Redo data validation for Tag* and Menu*
The menu adds options to consider, but we still want to make
sure we are testing on the basics
(arbitrary/duplicates/allowed list/etc). Make sure that the
isAllowedData method is running properly in the base class
and also in the Menu class.

Also, add unit tests for data validation.

Change-Id: Id32f37d9bf1828cae616effabf3c2377697cd6fe