Do not use role=menu/menuitem for MenuSelectWidget/MenuOptionWidget
They will now use role=listbox/option from SelectWidget/OptionWidget.

The difference is a bit unclear, but I think listbox is more
appropriate. Per https://www.w3.org/TR/wai-aria/roles:
"A menu is often a list of common actions or functions that the user
can invoke. The menu role is appropriate when a list of menu items is
presented in a manner similar to a menu on a desktop application."
Usually, MenuSelectWidget items do not correspond to user actions
(they are just values), and it's not really a desktop-like menu
(that sounds more like our ListToolGroup and other toolbar stuff).

And for a more practical consideration: a role=combobox widget (like
our ComboBoxInputWidget) can only be paired with a role=listbox list,
not with a role=menu one.

Bug: T149654
Change-Id: Ib9894ef9c70132e1d820bfa3815e47e5160d7ecd