SelectWidget/MenuSelectWidget: Maintain `aria-activedescendant` attribute on focus owner
Our SelectWidget does not allow individual OptionWidgets to have
focus. Instead, it keeps track internally of which of the options
really is active (e.g. which option arrow navigation is relative to,
or which option will be selected if the user presses the enter key).
This makes some implementation details a bit easier.

However, it means we have to somehow inform assistive technology about
the currently active option, since they can't guess it from the focus.
This is done by setting `aria-activedescendant` on the DOM element
which actually has focus (the focus owner).

This is made a little more complicated because the SelectWidget can
stand on its own (e.g. RadioSelectWidget), in which case it is the
focus owner itself, or it can be a part of a larger widget (e.g.
MenuSelectWidget/DropdownInputWidget), in which case the focus owner
is usually somewhere else.

As a bonus, SelectWidget will now move focus to the focus owner when
one of its options is focussed. The only situation where this makes a
difference is when a MenuOptionWidget is focussed through its accesskey.

Bug: T149654
Change-Id: Ie993345e44ddb43dfbe2bebe0e12fd2bd9f5812e