NumberInputWidget: Remake as an actual TextInputWidget child
As the name of it suggests, it should be a descendant of InputWidget
and, more specifically, of TextInputWidget. This solves a bunch of
compatibility issues.

To demonstrate this, a demo was added with TagMultiselectWidget
using NumberInputWidget; in its previous version the NumberInputWidget
could not have been used inside the TagMultiselectWidget because it didn't
have the expected API of TextInputWidget.

Reasons to use TextInputWidget rather than InputWidget:
* Appearance/logic: NumberInputWidget has a textbox in it and basically
  looks and feels just like a TextInputWidget except it is numbers-only
  and has the up/down buttons. Appearance-wise, it makes sense to inherit
  the styles of TextInputWidget, and override specifics.
* Properties: TextInputWidget uses the same type input as NumberInputWidget
  which means that things like 'indicator' and 'icon' and 'placeholder'
  can be seamlessly used and be consistent with the other TextInputWidget
  children. For example, adding a "star" indicator to denote that the
  widget is required (like we do in PHP) is seamless if we use
  TextInputWidget, but if we use InputWidget, we'll require redoing the
  IndicatorElement and IconElement logic.
  Another example is 'read only' states, which refer to the $input prop
  and 'setRequired' which also does the same. Using InputWidget would
  mean reapplying these.
* Methods like 'setValidityFlag' change the attr of the $input, which is
  also very useful, coming from TextInputWidget
* Label: InputWidget is not LabeledElement (perhaps it should be) while
  TextInputWidget is.
* Events: TextInputWidget listens to relevant events from the input,
  like 'change', keypress, focus and blur events on the input - all
  of which make sense to use in the NumberInputWidget as well.

In general, the NumberInputWidget -- having its main functionality an
actual <input type="number"> that allows typing -- is similar enough to
TextInputWidget to inherit from it, rather than from the parent InputWidget.

There are a few configuration options that come from TextInputWidget and
either don't make sense for the NumberInputWidget or are redundant but
we can either cancel those out (override them in the $.extend( config...)
in the constructor) or discourage their use in general. Overall, though,
it seemed to make more sense to inherit the useful aspects of
TextInputWidget rather than re-implement them when inheriting from
InputWidget.

Bonus: Renamed the horribly named 'setIsInteger' to 'setAllowInteger'
(and the getter as well) and added backward compatibility (mostly
at James_F's behest)

Bug: T124856
Bug: T149644
Change-Id: I00164fcaf5092b60243d4e3cdc8992285cee33b4