Generalize .getInputId() for all widgets
No reason to limit this to InputWidget, the <label for> is nice for
accessibility so let's use it where we can. This method is now a part of
TabIndexedElement. Widget includes a fallback version of it, so that we
can just call it for anything.

The 'supportsSimpleLabel' property (previously meant to allow for the same
thing but broken at least since 73521855861d4ef887af0b67533245c15e4dd942,
which added a requirement for widgets to also implement .getInputId()),
is now removed.

The HTML specification places some arbitrary restrictions on which
elements a <label for> can be used with (so-called 'labelable elements'),
which TabIndexedElement follows.

Also resolve minor inconsistencies in the handling of this in LabelWidget
versus FieldLayout (basically, FieldLayout was able to handle some
non-InputWidgets like CapsuleMultiselectWidget, LabelWidget wasn't).

This should result in no behavior changes for code which previously worked.
For some widgets (e.g. TagMultiselectWidget), a FieldLayout will now use
a `<label for>` rather than manual click handling, but for a user the
effect is the same.

Change-Id: I4286ca7633e29fcef551cffd74d46aed52819f95