Re-introduce .simulateLabelClick() as a separate method from .focus()
This implements the behavior that we would want a `<label for>`
associated with the widget to have, if HTML allowed us to associate
`<label for>` with arbitrary HTML elements.

* For most focusable widgets, it just focuses the widget.
* For CheckboxInputWidget, RadioInputWidget and ToggleSwitchWidget,
  it also toggles the widget's state.

FieldLayout and LabelWidget now call .simulateLabelClick() instead of
.focus() for widgets that can't be used with `<label for>`.

We previously deprecated InputWidget#simulateLabelClick as unnecessary
(73521855861d4ef887af0b67533245c15e4dd942). But it turns out that
such a method would be useful on some other widgets. Let's reuse the name.

InputWidget#simulateLabelClick is removed. Calling it will fall back to
TabIndexedElement#simulateLabelClick, which effectively works the same.

Bug: T164507
Change-Id: I50eeb5a65a9cdc343f7c199f4f2447aa6b5d032c