demos: Fix code generation for more complicated cases
This was broken in the last patchset of Ib3f59a68eecc42ca90ce
(https://gerrit.wikimedia.org/r/#/c/353960/18..19/demos/demo.js),
which changed the indentation from 4 spaces to 1 tab, but did
not change the indentation regexp to match tabs instead of spaces.

Also re-add the removal of extra indentation from functions
(e.g. in the TextInputWidget 'validate' config option), now
with a comment to explain the regexp and actually working.

Bug: T165457
Change-Id: I386d477438e6b2dfc0a811704da32b142a9dc6be