Apex theme: Make OptionWidget icon override more specific
We're overriding option widget's selected state with a 'check' icon
but since we're doing that to any and all iconElements under the
option widget, it also picks up stray widgets that it should not.

The rule should only apply to the direct descendent iconElement-icon
of the widget so it doesn't override cases where the optionWidget
contains other icons or buttons.

Bug: T165444
Change-Id: Ibe4b4fc92de1a864059d9cd37168c7ef33b6d874