MenuSelectWidget: Fix item hiding when menu contents change
updateItemVisibility() adds a CSS class to hide the menu when
it hides all items, but that means it should also reevaluate whether
all items are hidden when new items are added. Lookup-style widgets
can add items to menus dynamically, and those items should immediately
be hidden if they don't match the search string, or cause the
menu to be unhidden if they do match.

Add updateItemVisibility() calls to addItems() and removeItems().
Also make updateItemVisibility() the function that should generally
be called when things change, which means making it safe to call
if there is no input or input filtering, and using it to clip
instead of calling clip() directly.

Change-Id: Iec3c6757c7ab5d8866dfb4d5769008a34fb52fb6