When TitledElement is an AccessKeyedElement, show access key in the title
We use `jquery.accessKeyLabel` to get the key combination with modifier
keys. If it's not loaded, we only show the access key itself without
modifiers. Perhaps we should reimplement that logic one day, or add a
real dependency.

In the future, we should probably also make the text appended to the
title localisable. Currently, we have no method to localize strings
in PHP code at all.

I'm unhappy that both TitledElement and AccessKeyedElement have to
check if the other mixin/trait is mixed/used. This could be solved a
bit more neatly with events, but again, we have no events in PHP code.

Bug: T168408
Change-Id: I5c86ee6b256f2955e09cb98130bda2ef3a982fde