TagMultiselectWidget: Skip updateInputSize() for invisible inputs
When the widget is not attached yet, or if it is inside a display:none;
container, measuring the scrollWidth will fail. If updateInputSize()
is called during this time (and it is called many times during
initialization, as well as from an rAF callback), it will
set the input width to 1em (which looks bad) and also incorrectly
cache contentWidthWithPlaceholder as 0 (which breaks future resizes).

Bug: T169384
Change-Id: I08b691950adc78c9c58d85a78a705f4a3ea770b8