TextInputWidget: When positioning label, don't clear padding if we will set it again
This in most cases prevents the unpleasant "flickering" effect, and
the text cursor being hidden under the label, when typing at the end
of a horizontally scrollable input with a dynamic label.

However, if the width of the label increases, the issue still occurs.
This case needs some further thought.

Bug: T169982
Change-Id: I06d539c63645fe8698a73ac1749b5d3ebbba6ede