Follow-up 022f532: don't crash if TitledElement initializes before AccessKeyedElement
If a widget mixed in both TitledElement and AccessKeyedElement,
and called the TitledElement constructor first, then that would
call updateTitle() which calls formatTitleWithAccessKey(),
which then tries to access this.$accessKeyed before it's initialized.
This led to a JS error, but only if $.fn.updateTooltipAccessKeys
was set.

To prevent this from happening, short-circuit formatTitleWithAccessKey()
if this.$accessKeyed is not yet set, and call updateTitle() in the
AccessKeyedElement constructor to make sure it runs again.

Change-Id: I6c212c97da350554c9f364b4e967b0c6dea31d43