Support 'expanded: false' for MenuLayout, BookletLayout, IndexLayout
This allows them to behave sensibly when placed in a container that
doesn't have a fixed height. In particular, it is now much easier to
use them directly on a page, rather than inside OOUI dialogs.

The DOM order is changed to place the menu before the content (except
for `menuPosition: 'bottom'`. This will affect the tab order of the
fields, but should not cause visual changes.

Bug: T98141
Change-Id: I19cd1f2f3f7c4afa22a507a89c75fb206386eb0d