[DEPRECATING CHANGE]: Apex theme: Begin killing `constructive` flag
We chose to remove this flag from the WikimediaUI theme (then called
MediaWikiUI) in 89880581 as we decided that it added too much further
complexity without enough user value, and to reduce the payload cost
in terms of download size. We did this by making the `constructive`
flag equivalent to the `progressive` one.

We're now also removing it from the Apex theme, and also from OOjs UI
entirely. A follow-up breaking change will drop the fall-back values,
until which time the payload reduction will be minimal.

Bug: T156874
Change-Id: If83ab721bd75393db4c7dbbaf2a1749b82a6329e