Downgrade 'grunt-exec' to 1.0.1
The later versions are completely broken on Windows.

Version 2.0 of grunt-exec is a complete rewrite (for no documented
reason) to split the given command line into the command and arguments
and use child_process.spawn() instead of child_process.exec().

Unfortunately the splitting code is broken (e.g. doesn't handle '&&',
instead passes it as an argument to the first command) and the
spawning code is apparently also broken (trying to spawn 'composer',
which resolves to 'composer.bat', behaves as if the file didn't
exist).

Version 3.0 is an avoidable followup release because in addition to
the issues above something about quote parsing was also broken.

These issues are not visible on other operating systems because
apparently the developer noticed the brokenness on his machine and
special-cased non-Windows systems to not do the broken parsing.
Presumably the behavior on Windows was never tested.

There is no changelog for any of these changes.

I think we should never update this package again, and evaluate
switching to a different one. It seems to have been taken over by a
different developer than the original one and he doesn't know what
he's doing.

Change-Id: Ie958df88affff51af377d518a4364c66ed61d2fe