Fix .oo-ui-selectable() mixin to actually undo .oo-ui-unselectable()
The default value that we should be restoring is 'text', not 'all'.
(Technically, the default value is 'auto', but that seems to cause the
value to be inherited, so that doesn't work for us.)

Using 'user-select: all' only allows selecting the *entire* element,
rather than allowing normal text selecting. (It also doesn't seem to
actually work in browsers other than Firefox. MDN documentation hints
that it's only supposed to work in editable documents.)

The only place where we use this is to allow selecting contents of
popups in toolbars (toolbars are marked as unselectable).

Bug: T178320
Change-Id: Id81b36fb4ec5d39c07af9b920e59abdc18b889f7