MenuSelectWidget, PopupWidget: Automatically change popup direction if there is no space
The general idea of "flip direction if the default direction causes
clipping" is lifted from the implementation in PopupToolGroup, but the
details are different.

Note that we do not flip the direction if the clipping starts after
the widget is already visible (e.g. when the user scrolls), that seems
like it would be annoying.

PopupWidget gains a new config option 'autoFlip' (default 'true'),
which allows this behavior to be disabled. It's always enabled for
MenuSelectWidget.

Bug: T158934
Change-Id: Ifa26336b501c8aac0aebc8f9fe1120adc4b6c8bf