Allow adding virtual viewport spacing
Users can now override OO.ui.getViewportSpacing() to return additional
spacing that should be taken into account when displaying elements
that are clipped to the viewport, e.g. dropdown menus and popups.
This is meant to be overridden to avoid such menus overlapping any
fixed headers/toolbars/navigation used by the site.

Various code dealing with positioning/clipping elements now makes use
of that functions to pretend that the browser viewport is smaller than
it really is.

Note that funny things will happen if you override it and then put
OOUI widgets inside such a menu, as the widgets will always be
considered to be outside of viewport.

Bug: T106095
Change-Id: Iabfc1ea50f4eb3239e27d87e404ad13e1cddc3bb