DropdownWidget: Allow pressing Space to close the widget, as well as open
For any other widget with keyboard handling we treat Space and Enter
the same. This was an exception because Space conflicted with
type-to-search (see 2650383e), but I think I managed to solve that.

(The old code was additionally problematic because since 7e442e09
we allow type-to-search even when the menu is closed, so it would
suddenly pop open when you typed a space.)

Change-Id: I44b98c63d026630350c975eeb5ef2616c3b31717