DropdownInputWidget: Generate a hidden `<select>` in JS
This removes a big inconsistency between normal DropdownInputWidget
and infused DropdownInputWidget (where we kept the PHP `<select>`).

The `<select>` is currently always hidden, but we could easily decide
to display it; see T180730 for reasons to maybe do so.

The selected value of the `<select>` and of our pretty DropdownWidget
is synchronised in both directions. This actually didn't require any
changes (InputWidget and DropdownInputWidget#setValue already do it),
but I'm hereby documenting it as a feature.

Unblacklist DropdownInputWidget for JS/PHP comparison tests.
Blacklist or ignore some specific stuff instead.

Bug: T180730
Change-Id: Ie616726c158af0aa2f90e794f8d2ed3fd1e35cb5