Fix popup direction changing when the "anchor" is partially offscreen
When the "anchor" (floatable container) is partially offscreen, and
'hideWhenOutOfView' is not disabled, the popup/menu (floatable) gets
hidden. That means we get a 0x0px box when calculating its dimensions,
which in turn means it never has to be clipped.

So, whenever we're checking whether the floatable is clipped, let's
also check if it's completely offscreen.

Bug: T182650
Change-Id: I3568f445d5354c6bedee2c532c538b5e8b64ebe2