Avoid having to call .setValue() in some widgets' constructors multiple times
We had three InputWidget subclasses that had to call .setValue() for
the second time in their constructor, rather than rely on InputWidget
to do this: DropdownInputWidget, RadioSelectInputWidget and
CheckboxMultiselectInputWidget.

For example, the flow for DropdownInputWidget looked like this:

- DropdownInputWidget#(construct)
  - InputWidget#(construct)
    - DropdownInputWidget#setValue
      - (check options: but no options are defined yet, so this rejects all values)
  - DropdownInputWidget#setOptions
    - DropdownInputWidget#setValue [in case current value was removed from the options]
      - InputWidget#setValue
        - (accesses this.$input; this call can't be moved before InputWidget constructor)
  - DropdownInputWidget#setValue
    - (check options: now we can correctly check the validity)

To resolve this, we split #setOptions into two halves. A new method
named #setOptionsData is called before the InputWidget constructor
and carefully avoids touching the uninitialized properties. Updating
them is done after/during the InputWidget constructor from #setValue
(and #updateOptionsInterface). #setOptions simply calls both of these.

The new flow for DropdownInputWidget looks like this:

- DropdownInputWidget#(construct)
  - DropdownInputWidget#setOptionsData
  - InputWidget#(construct)
    - DropdownInputWidget#setValue
      - (check options)
      - DropdownInputWidget#updateOptionsInterface
      - InputWidget#setValue

Overall, this change should be a minor performance improvement, plus
it makes it easier to preserve the "original" value, which we need for
T181691.

Change-Id: Ida5a6d9a961f746c6431796211203e99158ce788